refactor: confirmWinningInfo 디렉토리를 confirmWinningResult로 변경
1. 디렉토리 네이밍 변경

2. 변경 사항 다른 모듈들에 적용
