+++ src/domain/confirmWinningResult/index.js
