+++ src/domain/confirmWinningResult/rateOfReturnCalculation/RateOfReturnCalculation.test.js
