+++ src/domain/confirmWinningResult/rateOfReturnCalculation/rateOfReturnCalculation.module.js
