+++ src/domain/confirmWinningResult/winningInfoGeneration/WinningInfoGeneration.test.js
