+++ src/domain/confirmWinningResult/winningInfoGeneration/winningInfoGeneration.module.js
